Namespace to module
